+++ frontend/src/components/Charts/DepthChart/index.tsx
@@ -23,12 +23,10 @@ import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { Order, LimitList } from '../../../models';
 import RobotAvatar from '../../RobotAvatar';
-import { amountToString } from '../../../utils/prettyNumbers';
+import { amountToString, matchMedian, statusBadgeColor } from '../../../utils';
 import currencyDict from '../../../../static/assets/currencies.json';
 import { PaymentStringAsIcons } from '../../PaymentMethods';
 import getNivoScheme from '../NivoScheme';
-import median from '../../../utils/match';
-import statusBadgeColor from '../../../utils/statusBadgeColor';
 
 interface DepthChartProps {
   orders: Order[];
@@ -93,7 +91,7 @@ const DepthChart: React.FC<DepthChartProps> = ({
     if (xType === 'base_amount') {
       const prices: number[] = enrichedOrders.map((order) => order?.base_amount || 0);
 
-      const medianValue = ~~median(prices);
+      const medianValue = ~~matchMedian(prices);
       const maxValue = prices.sort((a, b) => b - a).slice(0, 1)[0] || 1500;
       const maxRange = maxValue - medianValue;
       const rangeSteps = maxRange / 10;
@@ -104,7 +102,7 @@ const DepthChart: React.FC<DepthChartProps> = ({
     } else {
       if (lastDayPremium === undefined) {
         const premiums: number[] = enrichedOrders.map((order) => order?.premium || 0);
-        setCenter(~~median(premiums));
+        setCenter(~~matchMedian(premiums));
       } else {
         setCenter(lastDayPremium);
       }
