+++ frontend/src/components/Dialogs/AuditPGP.tsx
@@ -14,7 +14,7 @@ import {
   Link,
 } from '@mui/material';
 
-import { saveAsJson } from '../../utils/saveFile';
+import { saveAsJson } from '../../utils';
 import { systemClient } from '../../services/System';
 
 // Icons
