+++ frontend/src/utils/match.ts
@@ -1,7 +1,7 @@
-export const median = (arr: number[]) => {
+export const matchMedian = (arr: number[]) => {
   const mid = Math.floor(arr.length / 2);
   const nums = [...arr].sort((a, b) => a - b);
   return arr.length % 2 !== 0 ? nums[mid] : (nums[mid - 1] + nums[mid]) / 2;
 };
 
-export default median;
+export default matchMedian;
